Refactoring contact part
