+++ src/main/resources/lib/types/partTypes/contact.ts
@@ -1,10 +1,12 @@
-import { type Contact } from '/site/content-types/contact'
-
 export interface ContactModel {
   contactTitle: string
-  contacts: Array<Array<TransformedContact>>
+  contacts: Array<Array<Contact>>
 }
 
-export interface TransformedContact extends Contact {
-  phonelink: string
+export interface Contact {
+  id: number
+  name: string
+  email?: string
+  phone?: string
+  phoneLink?: string
 }
