+++ src/main/resources/site/parts/contact/contact.html
@@ -2,12 +2,20 @@
   <h2 data-th-text="${label}?: ${contactTitle}" class="roboto-condensed-bold mb-md-4 pb-md-3 bt-6"></h2>
   <div data-th-if="${contacts}" data-th-each="contactGroup : ${contacts}" class="row d-flex flex-wrap">
     <div data-th-if="${contact}" data-th-each="contact : ${contactGroup}" class="col-md-4 col-lg-3 contacts">
-      <p class="roboto-plain lead mb-1 text-break text-wrap part-contact-name" data-th-text=${contact.name} />
+      <p class="roboto-plain lead mb-1 text-break text-wrap part-contact-name" data-th-text="${contact.name}" />
       <p data-th-if="${contact.email}" class="part-contact-email mb-0">
-        <a class="roboto-plain ssb-link stand-alone" data-th-href="@{'mailto:' + ${contact.email}}" data-th-text="${contact.email}" />
+        <a
+          class="roboto-plain ssb-link stand-alone"
+          data-th-href="@{'mailto:' + ${contact.email}}"
+          data-th-text="${contact.email}"
+        />
       </p>
-      <p data-th-if="${contact.telephone} != ''" class="part-contact-telephone">
-        <a class="ssb-link stand-alone" data-th-text="${contact.telephone}" data-th-href="@{'tel:' + ${contact.phonelink}}" />
+      <p data-th-if="${contact.phone} != ''" class="part-contact-telephone">
+        <a
+          class="ssb-link stand-alone"
+          data-th-text="${contact.phone}"
+          data-th-href="@{'tel:' + ${contact.phonelink}}"
+        />
       </p>
     </div>
   </div>
