+++ src/main/resources/site/parts/contact/contact.ts
@@ -2,14 +2,13 @@ import { type Content } from '/lib/xp/content'
 import { getContent, getComponent } from '/lib/xp/portal'
 import { render } from '/lib/thymeleaf'
 import { type Phrases } from '/lib/types/language'
-import { type Contact } from '/lib/ssb/dashboard/statreg/types'
-import { find } from '/lib/vendor/ramda'
-
+import { type Contact as StatRegContacts } from '/lib/ssb/dashboard/statreg/types'
 import { renderError } from '/lib/ssb/error/error'
 import { getContactsFromRepo } from '/lib/ssb/statreg/contacts'
+import { getSelectedContacts } from '/lib/ssb/parts/contact'
 import { ensureArray, chunkArray } from '/lib/ssb/utils/arrayUtils'
 import { getPhrases } from '/lib/ssb/utils/language'
-import { type ContactModel, type TransformedContact } from '/lib/types/partTypes/contact'
+import { type ContactModel, type Contact } from '/lib/types/partTypes/contact'
 import { type Article, type Statistics } from '/site/content-types'
 
 const view = resolve('./contact.html')
@@ -26,30 +25,6 @@ export function preview(req: XP.Request) {
   return renderPart(req)
 }
 
-// split 8-digit phone numbers into groups of 2 digits each dvs. "12345678" => "12 34 56 78"
-function treatPhoneNumber(phone: string): string {
-  const matcher = /..?/g
-  const phoneArr: Array<string> | null = phone.match(matcher)
-  if (phoneArr) {
-    return phoneArr.join(' ')
-  } else {
-    return ''
-  }
-}
-const landCodeVisual = '(+47) '
-const landCode = '+47'
-
-function transformContact(contact: Contact, language: string): TransformedContact {
-  return {
-    ...contact,
-    telephone:
-      language == 'en' && contact.telephone != ''
-        ? landCodeVisual.concat(treatPhoneNumber(contact.telephone as string))
-        : treatPhoneNumber(contact.telephone as string),
-    phonelink: landCode.concat(contact.telephone as string),
-  }
-}
-
 function renderPart(req: XP.Request): XP.Response {
   const WIDTH = 4 // how many boxes in a row
   const page = getContent<Content<Article | Statistics>>()
@@ -61,7 +36,7 @@ function renderPart(req: XP.Request): XP.Response {
 
   const phrases: Phrases = getPhrases(page) as Phrases
 
-  const statRegContacts: Array<Contact> = getContactsFromRepo()
+  const statRegContacts: Array<StatRegContacts> = getContactsFromRepo()
   let contactIds: Array<string> = []
 
   if (part.config.contacts) {
@@ -71,18 +46,8 @@ function renderPart(req: XP.Request): XP.Response {
     contactIds = contactIds.concat(ensureArray(page.data.contacts))
   }
 
-  const selectedContacts: Array<TransformedContact> = contactIds.reduce((acc: Array<TransformedContact>, contactId) => {
-    const found: Contact | undefined = statRegContacts
-      ? find((contact: Contact) => `${contact.id}` === `${contactId}`)(statRegContacts)
-      : undefined
-    if (found) {
-      return acc.concat(transformContact(found, pageLanguage))
-    } else {
-      return acc
-    }
-  }, [])
-
-  const contacts: Array<Array<TransformedContact>> = chunkArray(selectedContacts, WIDTH)
+  const selectedContacts = getSelectedContacts(contactIds, statRegContacts, pageLanguage)
+  const contacts: Array<Array<Contact>> = chunkArray(selectedContacts, WIDTH)
 
   const contactTitle: string = phrases.contact
   if (!contacts || contacts.length < 1) {
